LA-91 Move the command modules to a new folder to make context clearer and root folder cleaner.
