+++ ArticleShapeExtractor/commands/CreateLeadArticle.jsx
@@ -1,8 +1,8 @@
 CreateLeadArticle = function() {
     this.run = function() {
-        require('./bootstrap.jsx');
+        require('../bootstrap.jsx');
         try {
-            const Container = require("./modules/Container.inc.jsx");
+            const Container = require("../modules/Container.inc.jsx");
             Container.resolve("InDesignArticleService").addOrRenameInDesignArticle("Lead");
         } catch (error) {
             error.alert();
