+++ ArticleShapeExtractor/commands/CreateSecondaryArticle.jsx
@@ -1,8 +1,8 @@
 CreateSecondaryArticle = function() {
     this.run = function() {
-        require('./bootstrap.jsx');
+        require('../bootstrap.jsx');
         try {
-            const Container = require("./modules/Container.inc.jsx");
+            const Container = require("../modules/Container.inc.jsx");
             Container.resolve("InDesignArticleService").addOrRenameInDesignArticle("Secondary");
         } catch (error) {
             error.alert();
